% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popreweigh.R
\name{reweigh}
\alias{reweigh}
\title{Finds new weights that are as close as possible to the original weights to
achieve the desired prevalence}
\usage{
reweigh(
  F,
  w,
  p,
  lambda = 1000L,
  lb_weights = 0,
  ub_weights = 1,
  lb_errors = -1,
  ub_errors = 1,
  scaling = NULL,
  weight_scale = NULL,
  solver = c("gurobi", "highs"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{F}{A numeric feature matrix (n x k) or a data.frame. If a data.frame,
it may contain the weight column specified by \code{w}.}

\item{w}{A numeric vector of original weights (length n). If \code{F} is a
data.frame, \code{w} can be the name of the weight column in \code{F}.}

\item{p}{A numeric vector of desired prevalences (length k).}

\item{lambda}{The penalty on errors (a scalar or a vector of length k).}

\item{lb_weights}{The lower bound for the weights (a scalar or a vector of
length n).}

\item{ub_weights}{The upper bound for the weights (a scalar or a vector of
length n).}

\item{lb_errors}{The lower bound for the errors (a scalar or a vector of
length k).}

\item{ub_errors}{The upper bound for the errors (a scalar or a vector of
length k).}

\item{scaling}{The scaling method for weights and features. One of "none",
"weights", or "auto". See \code{\link[=solver_model]{solver_model()}} for details.}

\item{weight_scale}{The scaling factor for the weights.}

\item{solver}{The solver to use, either "gurobi" or "highs". The solver must
be installed.}

\item{verbose}{If \code{TRUE}, solver output is printed. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to the solver.}
}
\value{
A list with the following elements:
\itemize{
\item \code{status}: The optimization status.
\item \code{weights}: A numeric vector of new weights (length n).
\item \code{prevalence_error}: A numeric vector of prevalence errors (length k).
\item \code{objective_value}: The numeric value of the objective function.
\item \code{solver}: The original solver result.
}
}
\description{
Finds new weights that are as close as possible to the original weights to
achieve the desired prevalence
}
