% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popreweigh.R
\name{reweigh_fem}
\alias{reweigh_fem}
\title{Reweigh a donor pool to match target prevalences}
\usage{
reweigh_fem(
  donors,
  population = 1,
  weight_var = "weight",
  original_prevalences = NULL,
  target_prevalences = NULL,
  trends = NULL,
  solver = c("gurobi", "highs"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{donors}{A data.frame representing the donor pool. It must contain the
\code{weight_var} and the variables for which prevalences are to be matched.}

\item{population}{The target population size. Default is 1.}

\item{weight_var}{A character string specifying the name of the column in
\code{donors} that contains the weights. Defaults to "weight".}

\item{original_prevalences}{A named numeric vector of original prevalences.
The names must correspond to columns in \code{donors}. If \code{NULL} (the default),
the original prevalences are calculated from the \code{donors} data.frame.}

\item{target_prevalences}{A named numeric vector of target prevalences. The
names must correspond to columns in \code{donors}. Either \code{target_prevalences}
or \code{trends} must be provided.}

\item{trends}{A named numeric vector of trends. Trends are multipliers
applied to the \code{original_prevalences} to obtain the \code{target_prevalences}.
Either \code{target_prevalences} or \code{trends} must be provided.}

\item{solver}{A character string specifying the solver to be used for the
reweighing optimization. Defaults to "gurobi".}

\item{verbose}{A logical value indicating whether to print the prevalence
errors after reweighing. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to the \code{reweigh()} function.}
}
\value{
A data.frame identical to the input \code{donors} but with the
\code{weight_var} column updated with the new weights.
}
\description{
This function reweighs a donor pool to match target prevalences for a set of
variables. The target prevalences can be specified directly or as trends
applied to original prevalences.
}
