% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popreweigh.R
\name{reweigh_gurobi}
\alias{reweigh_gurobi}
\title{Reweighs with Gurobi}
\usage{
reweigh_gurobi(model, verbose, ...)
}
\arguments{
\item{model}{The model created by \code{solver_model}.}

\item{verbose}{If \code{TRUE}, solver output is printed.}

\item{...}{Additional parameters for the Gurobi solver.}
}
\value{
A list with the optimization results.
}
\description{
Reweighs with Gurobi
}
\keyword{internal}
