% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popreweigh.R
\name{reweigh_highs}
\alias{reweigh_highs}
\title{Reweighs with HiGHS}
\usage{
reweigh_highs(model, verbose, ...)
}
\arguments{
\item{model}{The model created by \code{solver_model}.}

\item{verbose}{If \code{TRUE}, solver output is printed.}

\item{...}{Additional parameters for the HiGHS solver.}
}
\value{
A list with the optimization results.
}
\description{
Reweighs with HiGHS
}
\keyword{internal}
