% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popreweigh.R
\name{solver_model}
\alias{solver_model}
\title{Creates the solver model for reweighing}
\usage{
solver_model(
  F,
  w,
  p,
  lambda = 1000L,
  lb_weights = 0,
  ub_weights = 1,
  lb_errors = -1,
  ub_errors = 1,
  scaling = c("none", "weights", "auto"),
  weight_scale = NULL
)
}
\arguments{
\item{F}{numeric feature matrix (n x k) where n is the number of observations
and k is the number of features where n >> p.}

\item{w}{numeric vector of original weights (length n).}

\item{p}{numeric vector of desired prevalence (length k); if the features are
binary, p must be in interval between zero to one including; if the features
are continuous, p may be any real number; in this case, it is not prevalence
but weighted mean.}

\item{lambda}{penalty on errors (scalar or vector of length k).}

\item{lb_weights}{lower bound for the weights (scalar of vector of length n).}

\item{ub_weights}{upper bound for the weights (scalar of vector of length n).}

\item{lb_errors}{lower bound for the errors (scalar or vector of length k).}

\item{ub_errors}{upper bound for the errors (scalar or vector of length k).}

\item{scaling}{scaling method for weights and features; one of "none",
"weights", or "auto"; see details.}

\item{weight_scale}{scaling factor for the weights (scalar).}
}
\value{
model specification for cohort reweighing.
}
\description{
Creates the solver model for reweighing
}
\details{
\itemize{
\item scaling
\itemize{
\item equal to "none" means no scaling;
\item equal to "weights" means that weights are scaled so that their
average is equal to weight_scale; features are not scaled;
\item equal to "auto" means that weights are scaled so that the average
weight is equal to weight_scale if given or to 1 if n <= 1000 or to
sqrt(n / 1000) if n > 1000; features are standardized to have mean 0 and
standard deviation 1; p and bounds are scaled accordingly
}
\item if a feature is continuous instead of binary, its corresponding prevalence
is not a prevalence but its weighted mean; increased bounds may be necessary
\item BEWARE: many continuous features may require big changes in weights;
some weights may become huge, i.e., a great part of the new cohort may be
represented by few individuals and the error in prevalence may still be huge
}
}
\section{Typical use}{

\itemize{
\item for binary features:
\code{solver_model(F, W, p)} or
\code{solver_model(F, W, p, scaling = "weights")} when n is large
\item for mix of binary and continuous features or continuous features with very
different means/ranges:
\code{solver_model(F, W, p, scaling = "auto")}
\item with many continuous features, lower lb_errors and increase ub_errors but
be careful
\item if several individuals have too high weights, lower ub_weights
\item if no solution is found, lower lb_errors and increase ub_errors
}
}

\keyword{internal}
